// apps/chataigne/ui/src/lib/systems/state_machine/generated/FormulaPreviewModeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ContextKeyDto } from "./ContextKeyDto";

export type FormulaPreviewModeDto = { "kind": "formula_defaults", formula_id: string, } | { "kind": "processor_default_lane", processor_id: string, } | { "kind": "processor_lane", processor_id: string, context_key: ContextKeyDto, } | { "kind": "processor_inspection", processor_id: string, } | { "kind": "processor_selected_stages", processor_id: string, context_key: ContextKeyDto | null, node_ids: Array<string>, };